pyproject.toml
[project]
name = "project-5"
version = "0.1.0"
description = ""
authors = [
    {name = "Lenar Khaidarov",email = "haidarov_lenar@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.8,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "django-filter (>=25.2,<26.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "ipython (>=9.7.0,<10.0.0)",
    "coverage (>=7.12.0,<8.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.lint.dependencies]
mypy = "^1.18.2"
flake8 = "^7.3.0"
black = "^25.11.0"
isort = "^7.0.0"

[tool.black]
line_length = 119
extend-exclude = '''
(
  \.git
)
'''

[tool.isort]
line_length = 119

[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
exclude = '''
(
  \venv
)
'''
